Agent: Support complex place matching
If one part in the spec matches with multiple parts in the exported
place, then it's no longer a simple assignement. Add an implementation
of the Hopcroft-Karp algorithm, to solve this assignment problem.

Such cases may not occur that often, but when they do, it's nice to have
a solution.