Tests: Kill remote processes over SSH
The SSH client doesn't forward SIGTERM, that means if the remote process
doesn't use the stdio streams, it never notices, that it should
terminate. Fixed by watching stdin in a simple wrapper script
(while-stdin) and killing the process, when stdin is closed. The SSH
client isn't killed, but the tests wait for the remote process to exit,
so we don't have a race between cleanup of the old test and the start of
the next test.