JSON-RPC: Support cancellation
Raise a CancelledError on the remote side whenever a proxy method call
is canceled. This is especially useful for blocking calls, like
Server.reserve(): If the request is queued and the user decides to
cancel the request, then the request will be removed from the queue on
the server side.

The rpc.cancel method isn't defined in the JSON-RPC 2.0 specification,
and "rpc." is a reserved namespace. However it looks like the spec is
not maintained anymore and there were discussions about adding this
method to the spec:
https://groups.google.com/g/json-rpc/c/4uGcjO43yz0/m/nXu42GUrBwAJ