Replace hatch with meson-python
hatch is pretty slow when installing packages. Since the tests need to
install the package on every test run and all the three VMs, the
installation time is quite noticeable. Test runs on my machine are now
done in 15 seconds, instead of 23 seconds.

Another reason to replace hatch with an actual build system is, that I
want to generate and install man pages / documentation during the build.
I haven't found a way to do this with hatch.