USB/IP: Bind usbip-host driver in exporter
Initially I wanted to run the exporter without root permissions, so I
bound the usbip-host driver in the uevent command. The uevent command
isn't triggered however, when the exporter starts, so the device might
still be bound to its default driver.

To fix this move the driver binding to the exporter. The uevent command
only notifies the exporter if a new device appears.

Additionally unbind the default driver in the exporter to handle the
case when starting the exporter the first time. This might break
sensible USB devices, but a USB device reset should fix it.