Agent: Support exporters with loopback host
If an Exporter and the Hub run on the same host and the Exporter
connects to the hub via localhost, then the host of the Place is the
loopback address for every agent. This prevents agents from connecting
if they are on a remote host.

Fixed by replacing the host of the Place with the host of the Hub, if
the host of the Place is the loopback address.

I considered replacing the host in `/api/v1/places` on the Hub, but then
clients would get different responses based on their source IP address,
which doesn't work well with HTTP caching, if it will ever be used.