Hub: Merge WebSocket endpoints again
With the bidirectional jsonrpc channel it's now possible to use one
endpoint again. This makes it easier to integrate authentication,
because then there's only one connection handler.

Drop the dummy WebSocket authentication. Since ID tokens need to be
refreshed, a one time authentication is not enough. Without
authentication, all endpoints are the same, even for login.