CLI: Use tabulate to Print Status Table
When adding new columns after status, one has to account for ANSI escape
sequences when aligning the table, because the status column values are
printed red or green. At this point maintaining this logic is not worth
it and I'd rather add another dependency to the project.

Use python-tabulate to format the table. This has the added benefit,
that the width of the table is never more than necessary.