doc: Add GitHub link
ReadTheDocs no longer injects repository information in conf.py as it
used to. Furo doesn't support the new format, yet, so no links are
displayed.

Also the link to source no longer links to GitHub, but instead links to
a copy of the source code under _static/. Disable the source code link
completely for now.