doc: Fix sidebar scrolling on iOS Safari
Safari on iOS dynamically displays the URL bar at the bottom. The
viewport height unit includes the space behind the URL bar. That means
if the URL bar is displayed, you cannot click on the last element of the
sidebar.

Fix by using the newly introduced dynamic viewport height unit instead.
This unit includes only what's visible at the time, so you can scroll to
the last element as expected even if the URL bar is shown.