Agent: Fix matching duplicate parts
With a duplicate part in the import description, the agent would match a
place with just a single part in it.

The reason is, that there's a shortcut in the agent to avoid using the
more resource intensive Hopcroft-Karp algorithm. The shortcut logic is
wrong for duplicate parts. It only checks if more than one exported part
matches with one imported part, but it doesn't check if the same
exported part matches with multiple imported parts.

Fix by removing the shortcut and just use the Hopcroft-Karp algorithm
every time. It's still fast enough and without a noticeable delay.